requirements.txt
google-genai[aiohttp]>=1.65.0
tavily-python>=0.5.0
pydantic>=2.0
python-dotenv>=1.0.0
gradio>=5.0
